# internal/assets/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice Management API
  description: API for managing invoices, categories, companies, and file uploads
  version: 1.0.0
  contact:
    name: RxTech Lab
    url: https://github.com/rxtech-lab/invoice-management
servers:
  - url: /api
    description: API server
...
paths:
...
  /api/upload:
    post:
      tags:
        - Upload
      summary: Upload file
      description: Uploads a file to S3-compatible storage
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    UploadResponse:
      type: object
      properties:
        key:
          type: string
          description: S3 object key
        download_url:
          type: string
          format: uri
          description: Download URL for the file
        filename:
          type: string
          description: Original filename
        size:
          type: integer
          description: File size in bytes
        content_type:
          type: string
          description: MIME type
...
security:
  - BearerAuth: []